Refactor: Extract RepoMetaData to own file
