+++ source/dsl/RepoMetaData.ts
@@ -1,0 +1,7 @@
+/** Key details about a repo */
+export interface RepoMetaData {
+  /** A path like "artsy/eigen" */
+  repoSlug: string
+  /** The ID for the pull/merge request "11" */
+  pullRequestID: string
+}
