+++ source/platforms/_tests/_bitbucket_cloud.test.ts
@@ -32,7 +32,7 @@ jest.mock("../bitbucket_cloud/BitBucketCloudAPI", () => {
 
 import { BitBucketCloud } from "../BitBucketCloud"
 import { BitBucketCloudAPI } from "../bitbucket_cloud/BitBucketCloudAPI"
-import { RepoMetaData } from "../../dsl/BitBucketServerDSL"
+import { RepoMetaData } from "../../dsl/RepoMetaData"
 
 describe("getPlatformReviewDSLRepresentation", () => {
   let bbs: BitBucketCloud
